requirements.txt
# AI Tracker - GitHub AI Project Tracker
# Requirements for AI前沿追踪系统

# Core dependencies
requests>=2.31.0
PyGithub>=2.1.1
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Code quality (per constitution)
ruff>=0.1.0
mypy>=1.7.0

# Optional: For better API handling
httpx>=0.25.0
